+++ CHANGELOG.md
@@ -23,6 +23,7 @@
   - `ES_INDEXING_SLOWLOG_SOURCE`: Indexing slowlog source limit (default `1000`)
 - [TD-3222] `structures` index alias can now be configured using the
   `ES_ALIAS_STRUCTURES` environment variable
+- [TD-2951] `profile_structure` permission
 
 ## [4.14.0] 2021-2-22